refactor: KEEP-367 move ABI arg parser to lib/abi and share with args-list
parseAbiFunctionArgs / coerceAbiArgValue lived inside the React module
action-config-renderer.tsx, which forced the unit test to import a
client component just to reach a pure helper. Move both into the
existing lib/abi/ module as parse-args.ts. The component file imports
from there; the test imports from there directly.

Have args-list-field.tsx use coerceAbiArgValue too. Its previous
inline coercion treated non-array objects as primitives and pushed
"[object Object]" into the values array via String({}) -- which would
have broken any future tuple-object args that landed there. Sharing
coerceAbiArgValue gives args-list tuple-object passthrough for free.
The existing "converts non-string values to strings" test still passes
since the helper is a strict superset of the prior behavior.
